fix(ios): move QuickActionsIconPatcher into AppDelegate and fix build errors
- Moved patcher class into AppDelegate.swift to avoid manual pbxproj registration
- Fixed non-optional UIApplicationShortcutIcon(systemImageName:) guard binding
